requirements.txt
Django
djangorestframework
gunicorn
pandas
scikit-learn
scipy
tqdm
